HumanEval/app/requirements.txt
langchain==0.2.15
langchain-community==0.2.15
langchain-core==0.2.37
langchain-openai==0.1.23
langchain-text-splitters==0.2.2
langchain-huggingface==0.0.3
openai==1.55.3
pandas==2.2.2
psycopg2-binary==2.9.9
pydantic==2.8.2
python-dotenv==1.0.1
streamlit==1.29.0
streamlit-modal==0.1.2
google-generativeai==0.7.2
faiss-cpu
transformers